# Preserve data integrity when merging rows