Merging sorted row groups sometimes leads to data corruption, even when merging rows from a single reader. I suspect that values in rows returned by `head()` are getting overwritten when new rows are read into the buffer. Calling `Clone()` on the return value of `head()` fixes the problem: https://github.com/segmentio/parquet-go/blob/510764ae9e8007c4974dd8a1dc42f63f4bbcac4e/merge.go#L232

I have also added a test case which consistently reproduces the bug. I am happy to add in the fix, but I am not entirely whether cloning each row would have performance implications.