@fpetkovski 2023-07-03
Looks related to what is discussed in https://github.com/segmentio/parquet-go/issues/475.